+++ point-service/src/main/java/com/oneonone/pointservice/presentation/dto/request/UpdatePointStatusRequest.java
@@ -1,4 +1,4 @@
-package com.oneonone.pointservice.presentation.request;
+package com.oneonone.pointservice.presentation.dto.request;
 
 import com.oneonone.pointservice.domain.enums.PointStatus;
 import lombok.Getter;
